Merge branch 989-property-validation into trunk
This contains the updated common.inputs.Parameter class that serves as a model for future implementations. All classes will slowly be updated to the new standard. 
This also includes a demonstration of the new display approach. All classes must eventually be subclasses of CossanObject.m and override the method getPropertyGroups.

Additionally this includes a first version of my code-coverage toll that creates a detailed html report about the tests that have been run.