Temporary fix for Dataseries
Dataseries behaves not correctly with struct2table. In MATLAB this function behaves as

struct -> cell -> matrix

cell to matrix calls cat(2, object)

This temporary fix allows to create a correct table when a structure contains a Dataseries by blocking the execution of cat, since when MATLAB detects an error calling cat it reverts to the default behavior and don't convert the cell array to matrix.