Fix plotregression of MetaModel for legacy code
plotregression for metamodels that contains both Input and SimulationData objects in calibration and/or validation (typical for the old way to train a metamodel) was crashing because the function was passing the input object to evaluate, e.g., as if it was the old apply. 
It is now fixed and passes the table with the inputs.