Update all files with new opencossan package
The unit tests are now in the test/unit folder. Only the Parameter test has currently been migrated. I will move the other tests in the folder gradually as they will probably still require a lot of fixing.

I updated OpenCossan to also add the mex files and matlab-code-coverage to the path.