Merge branch 1003-update-test-tools
Tests are now run in docker container. The mac address in the Jenkinsfile is needed for the MATLAB license to work. Since it is local, exposing it should not be to dangerous.
I hope to find a better way in the future. The test should also check out a fresh copy, but this takes too long currently. Simulataneous tests might still interfere because the same folder is mounted into the container.

Closes #1003